Update orderby support for Price Benchmarks
This updates the PriceBenchmarks controller to specify a supported map of orderby values and maps them to the appropriate fields in the database query. The default value is also set as a constant in the class which is then referenced in the PriceBenchmarksController class that defines the API endpoint.

Unit tests are added to test all the supported orderby values and ensure that the API behaves as expected.